added types component genericized pokemon list for reuse card -> pokemonitem
